requirements.txt
flask
twilio
google-cloud-speech
google-cloud-texttospeech
requests